buildsys: add tidy make target
As of writing there are 51 .pm or .pl files checked into git, using
hyperfine to compare the optimal balance between parallelism and
perltidy start-up time it showed letting xargs pass at max 4 files to
a single perltidy call results in the shortest time on my 20 vCPU core
VM.

While this might not be ideal for all repos, it is low enough to be an
OK trade off and we can use it as default, and specific odd repos
could also get slightly different configurations.

hyperfine Summary:

'git ls-files ':*.p[ml]'| xargs -n4 -P0 proxmox-perltidy' ran
  1.04 ± 0.10 times faster than 'git ls-files ':*.p[ml]'| xargs -n3 -P0 proxmox-perltidy'
  1.06 ± 0.10 times faster than 'git ls-files ':*.p[ml]'| xargs -n6 -P0 proxmox-perltidy'
  1.09 ± 0.10 times faster than 'git ls-files ':*.p[ml]'| xargs -n7 -P0 proxmox-perltidy'
  1.10 ± 0.11 times faster than 'git ls-files ':*.p[ml]'| xargs -n5 -P0 proxmox-perltidy'
  1.12 ± 0.13 times faster than 'git ls-files ':*.p[ml]'| xargs -n2 -P0 proxmox-perltidy'
  1.12 ± 0.10 times faster than 'git ls-files ':*.p[ml]'| xargs -n8 -P0 proxmox-perltidy'
  1.20 ± 0.11 times faster than 'git ls-files ':*.p[ml]'| xargs -n9 -P0 proxmox-perltidy'
  1.26 ± 0.11 times faster than 'git ls-files ':*.p[ml]'| xargs -n11 -P0 proxmox-perltidy'
  1.40 ± 0.13 times faster than 'git ls-files ':*.p[ml]'| xargs -n10 -P0 proxmox-perltidy'
  1.41 ± 0.14 times faster than 'git ls-files ':*.p[ml]'| xargs -n1 -P0 proxmox-perltidy'
  1.41 ± 0.13 times faster than 'git ls-files ':*.p[ml]'| xargs -n12 -P0 proxmox-perltidy'

For the 88 files in pve-manager passing 4 arguments is the second
fastest behind passing 9, which might be a repo-specific outlier
though.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>