auto-format code using perltidy with Proxmox style guide
using the new top-level `make tidy` target, which calls perltidy via
our wrapper to enforce the desired style.

Inflight patches need to be rebased, which is naturally not ideal, but
there isn't a way around it. Albeit, as most churn stems from the
change from mixed tab-space to just space for indentation, one can
avoid lots of issues by passing the `--ignore-space-change` option to
git am, or git show to view this patch for that matter.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>