network: add helper for checking if a interface is a ovs/linux bridge
There are several places throughout multiple packages where we check
for the type of a bridge, so introduce helpers to avoid reimplementing
the logic across all callsites. The existing helper for detecting
linux bridges in pve-network checked for the existence of the brif dir
instead of bridge. I've decided to stick with bridge here, in my tests
they both only existed for linux bridges so they should be equivalent.

Co-authored-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>