tools: dump logfile: fix result type in case the file cannot be opened
The second return value of dump_logfile should be a reference to an
array of hash references (one per log line), but in case the file
cannot be opened, it currently directly returns a reference to a hash.

As a result, the /nodes/{node}/tasks/{upid}/log endpoint response does
not match the schema in case the task log cannot be opened (for
example, does not exist). This was reported in enterprise support, as
it trips up external monitoring tools that query this endpoint.

To fix this, make sure to return an array.

This technically also affects the endpoints that query the Ceph,
replication, and firewall logs, though there it is more likely that
the logfile actually exists and can be opened.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20250717165006.106126-1-f.weber@proxmox.com