schema: support sizes with verbose suffixes {K, M, G, T}iB
The single-letter suffixes are ambiguous and especially in the context
of disks, the powers of ten are usually used. Proxmox VE uses
multiples of 1024 however.

This is in preparation to adapt format_size() to prefer the verbose
suffixes which is planned for Proxmox VE 9.

The single-letter suffixes stay supported for backwards-compatibility
for API/CLI users as well as parsing backup configurations. The
function is currently also used for parsing the 'target-size' Ceph
pool option (still restricted via schema to single-letter suffixes).

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250723115809.96282-2-f.ebner@proxmox.com