inotify/interfaces: use 'ip link' instead of /proc/net/dev
The function reading /etc/network/interfaces used /proc/net/dev to
determine pre-existing physical interfaces. Since the introduction of
altnames, /proc/net/dev returns insufficient information for
determining if an interface is already contained in /e/n/i, since it
does not include altnames.

The interfaces parser added all interfaces from /proc/net/dev as
configurable network devices. If altnames were used in the
configuration, then the same interface would be listed twice: once
with its 'real' name (from /proc/net/dev) and once with its altname
(from the interfaces file).

Adapt the interfaces file parser to utilize 'ip link' instead of
/proc/net/dev as its source of truth regarding configured network
devices. The parser now only adds a physical interface if neither its
'real' name nor any of its altnames are contained in
/etc/network/interfaces.

This patch makes the assumption that a interface is always referred to
by either its name or one of its altnames, but not mixed (e.g. the
stanza uses the 'real' name but bridge_ports uses the altname). In its
current state this will already lead to problems with validation,
since we do not consider altnames when looking up the availability of
bridge ports or VLAN parent devices for instance.

As long as our tooling is used, this assumption should always hold, it
can only be broken by manually editing the interfaces file.

There is one special case still where two entries exist for a single
network interface: If a pinning is generated by
proxmox-network-interface-pinning, but not yet applied. In this case,
the UI will show both network interfaces (nicX and ethX). It will also
show the pending changes that replace ethX with nicX. A possible
solution for improving this case could look at the existing .link
files, check if any not-yet-applied pin exists there and use the
pinned name as the 'real' name when parsing the interfaces file.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250723142106.235104-1-s.hanreich@proxmox.com
 [TL: run through perltidy]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>