JSONSchema: property strings: consistently handle default_key of type boolean
To be able to upgrade existing simple single-value schemas to more
complex property strings our JSON schema allows to define one of the
properties as "default key", this is then the key that will be used to
parse a standalone value into.

The handling of parsing this was inconsistent for boolean values, a
"legacy" standalone `value` string (relying on default_key) was not
passed through parse_boolean whereas a full `key=value` property
string was actually passed through parse_boolean, making the latter
more flexible than the former.

For example, given the following schema:
```
enabled => {
    type => 'boolean',
    default_key = 1,
},
```
The following values are accepted with either representation:
* `0` and `enabled=0`
* `1` and `enabled=1`

But the following got inconsistent accepted/rejected:
* `true` (fails) `enabled=true` (works)
* `false` (fails) `enabled=false` (works)

To fix this, use parse_boolean normalization also for the case when
the property denoted as default_key is passed as standalone value.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250912093919.1715010-1-d.csapak@proxmox.com
 [TL: reword commit message]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>