create dedicated PVE::UPID module
This is part of the long-term goal to split pve-common, and especially
the overarching PVE::Tools module. Even though we will focus on rust
code for new projects, the Perl side won't be switched overnight and
so will stay with us for quite a few years, thus it still makes sense
to improve the developer experience here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>