tests: add basic test system for JSON schema specific tests
While we got some indirect coverage through some other tests, this is
worth having dedicated ones.

Start out simple with tests for a recent fix for booleans as default
key, but this really should be extended, but it *is* normally easier
to extend if the existing base infra is there.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>