file: add file_read_last_line method
Complementing the existing file_read_first_line one. The
implementation tries to stay simple, seek to the end and read backward
in chunks until we find a newline. The last bit gets slightly
complicated as normally all text files are allowed to end with a
newline, and we do not want to treat the empty string after that
trailing newline as last line, so add a special case for when the last
byte in the buffer is a newline at the first (read) iteration and
trim that newline out.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>