upid: use new helpers from File module
The core code for creating a file exclusively, chown it and return the
FH and the one for reading the last line of a file are rather generic
file stuff and thus a better fit for the relatively new PVE::File
module.

As a side benefit we can now easier test the PVE::UPID module, as
mocking functions from CORE like sysread, close or sysseek, is
extremely hard to do correctly, as Test::MockModule won't work (even
though they silence an error for mocking CORE::GLOBAL it still doesn't
do anything) and overriding them in a BEGIN statement is only a real
option if one never wants to call the original method, and even then
it's rather a PITA.

This was actually one of the main reasons that triggered this work in
the first place.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>