systemd: add notify() helper
Add a pure Perl reimplementation of systemd's sd_notify() mechanism as
defined in systemd/sd-daemon.h based on the example implementations in
'man 3 sd_notify'.

The initial user of this helper is intended to be the pve-dbus-vmstate
service, so it can notify startup completion only once the
dbus-vmstate QEMU object is ready to be used.

EAGAIN is not checked for, because it does not occur for blocking
Unix domain sockets, see 'man 2 send'. The socket is a datagram
socket and a single datagram will be sent. If not all data can be
written, error out like the example code in 'man 3 sd_notify'.

Co-developed-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Co-developed-by: Wolfgang Bumiller <w.bumiller@proxmox.com>
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251029100646.27658-2-f.ebner@proxmox.com