JSON schema: make expose_credentials flag available for API endpoints
This flag is intended to be used in API calls that need access to the
credentials of the user. That is mainly required for API calls that
want to manually proxy to other nodes.

Current use cases work around the lack of such a mechanism by setting
the protected flag, which makes the API call get executed as
privileged daemon, which can create a fresh ticket, it works, but it's
a big hammer to wield. And in addition to the high privileges used
this workaround also causes and additional proxy request, causing
potential slow down.

Co-authored-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>