fix #7153: iproute2: consider bonds as physical bridge ports
The new check did not consider bonds, since ip_link_is_physical only
returns true for physical interfaces - but not bonds. Add a new
function to check for bonds explicitly and use it when checking for
bridge ports in activate_bridge_vlan.

This bug caused VMs with network devices with VLAN tags, on
non-vlan-aware bridges, to not get started / migrated.

Fixes: 057f62f7
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251216160513.360391-1-s.hanreich@proxmox.com