tests: sectionconfig: add comparison test structure
During development, I spotted a couple interesting "peculiarities"
when it comes to the schemas that PVE::SectionConfig generates.
I was initially intending to document these cases textually (as part
of the PVE::SectionConfig docs), but quickly realized that it's a
better idea to document these cases as unit / regression tests.

Therefore, set up a new test directory specifically for
PVE::SectionConfig and add `schema_comparison_test.pl`. The latter
contains a setup that allows us to write declarative tests for
SectionConfig in the form of nested packages:

- Each "test package" is based on the top-level `TestPackage`

- Inside such a test package, a PVE::SectionConfig plugin structure
  must be declared (i.e. base plugin and child plugins)

- The create and update schemas of each plugin structure inside a
  test package are then compared against their expected outputs,
  which are also declared inside the test package

- this is done for both unified and isolated modes without the need to
  define separate packages for either mode

In other words, to add a new test case, one has to only define a test
package, add a PVE::SectionConfig plugin structure inside of it, and
then declare the expected outputs for the schemas.

To run the tests, find all of the child packages of `TestPackage`
inside `main` (the test script) and run a subroutine for it that sets
up and tears down the plugin structure for unified and isolated mode.

Should a test fail, the contents of the gotten and expected schemas
are printed to the console in a human-readable format using
Data::Dumper, with some customized parameters. This should make it
rather easy to spot any discrepancies in case a test breaks.

Finally, add a basic test case checking that properties added by child
plugins are always optional, even if they're declared as non-optional,
both to document the first "interesting fact" about SectoinConfig and
also in order to actually show how test cases are added.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-3-m.carrara@proxmox.com